Implement built-in search Remove Sass

+++ src/Client/Pages/HistoryView.razor.cs
@@ -30,7 +30,7 @@ public partial class HistoryView : OfflineSupportComponent, IAsyncDisposable
     /// <summary>
     /// The requested page number.
     /// </summary>
-    [Parameter] public long? PageNumber { get; set; }
+    [Parameter] public int? PageNumber { get; set; }
 
     /// <summary>
     /// The requested page size.
@@ -276,14 +276,14 @@ public partial class HistoryView : OfflineSupportComponent, IAsyncDisposable
         if (Revisions?.HasNextPage == true)
         {
             CurrentPageNumber++;
-            PageNumber = (long)CurrentPageNumber;
+            PageNumber = (int)CurrentPageNumber;
             await RefreshAsync();
         }
     }
 
     private async Task OnPageNumberChangedAsync()
     {
-        PageNumber = (long)CurrentPageNumber;
+        PageNumber = (int)CurrentPageNumber;
         await RefreshAsync();
     }
